refactor: вынести компонент App в папку components
