+++ src/components/App.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import './App.css';
-import Header from '../Header/Header';
-import Main from '../Main/Main';
-import Footer from '../Footer/Footer';
-import CurrentUserContext from '../../contexts/CurrentUserContext';
+import Header from './Header/Header';
+import Main from './Main/Main';
+import Footer from './Footer/Footer';
+import CurrentUserContext from '../contexts/CurrentUserContext';
 
 function App() {
   const [loggedIn, setLoggedIn] = React.useState(true);
